src/interpretability/requirements.txt
annotated-types
attrs
biopython
boto3
botocore
certifi
charset-normalizer
cloudpickle
dill
docker
einops
filelock
fsspec
google-pasta
idna
importlib-metadata
Jinja2
jmespath
jsonschema
jsonschema-specifications
kipoiseq
logomaker
MarkupSafe
matplotlib
moods
MOODS-python
mpmath
multiprocess
natsort
networkx
numpy
packaging
pandas
pathos
pathvalidate
platformdirs
polars
pox
ppft
protobuf
psutil
pydantic
pydantic_core
pyfaidx
pymemesuite
pyranges
pysam
python-dateutil
pytz
PyYAML
referencing
requests
rpds-py
s3transfer
sagemaker
schema
seaborn
six
smdebug-rulesconfig
sympy
tblib
tiny-api-client
torch
tqdm
typing_extensions
tzdata
urllib3
zipp
